+++ Sources/VSPlayer/Subtitle/SubtitleDataSouce.swift
@@ -84,7 +84,7 @@ public class PlistCacheSubtitleDataSouce: CacheSubtitleDataSouce {
     public static let singleton = PlistCacheSubtitleDataSouce()
     public var infos = [any SubtitleInfo]()
     private let srtCacheInfoPath: String
-    // 因为plist不能保存URL
+    // Because plist can not store URL
     private var srtInfoCaches: [String: [String]]
     private init() {
         let cacheFolder = (NSTemporaryDirectory() as NSString).appendingPathComponent("VSSubtitleCache")